refactor: hooks の作成
